Removing the duplication
