+++ sudokuru/app/Components/SudokuBoard/Core/Components/EndGameModal.tsx
@@ -1,10 +1,10 @@
-import { Button, List, Text, TouchableRipple } from "react-native-paper";
+import { Button, Text } from "react-native-paper";
 import { useWindowDimensions, ScrollView, View } from "react-native";
 import { useNavigation } from "@react-navigation/native";
 import Statistic from "../../../Statistics/Statistic";
 import { formatTime } from "../Functions/BoardFunctions";
 import React from "react";
-import { NumHintsUsedPerStrategy } from "../../../NumHintsUsedPerStrategy";
+import ExpandableHintsUsedStatistic from "../../../Statistics/ExpandableHintsUsedStatistic";
 import { useTheme } from "../../../../Contexts/ThemeContext";
 import { ClassicGameStatistics } from "../../../../Functions/LocalDatabase";
 
@@ -16,17 +16,10 @@ export const EndGameModal = ({
   const { theme } = useTheme();
   const size = useWindowDimensions();
   const reSize = Math.min(size.width, size.height);
-  const isLargeScreen = size.width >= 800;
-  const statFontSize = Math.max(
-    16,
-    Math.min(reSize / (isLargeScreen ? 19 : 23), isLargeScreen ? 30 : 24),
-  );
   const resultsMaxWidth =
     size.width >= 1100 ? 540 : size.width >= 800 ? 500 : 460;
 
   const navigation: any = useNavigation();
-  const [isHintsBreakdownExpanded, setHintsBreakdownExpanded] =
-    React.useState(false);
 
   return (
     <View
@@ -84,58 +77,10 @@ export const EndGameModal = ({
               statisticValue={statistics.difficulty}
               testID="difficulty"
             />
-            <TouchableRipple
-              onPress={() =>
-                setHintsBreakdownExpanded(!isHintsBreakdownExpanded)
-              }
-              style={{ marginBottom: 8 }}
-              rippleColor={theme.colors.border}
-            >
-              <View
-                style={{
-                  flexDirection: "row",
-                  justifyContent: "flex-start",
-                  alignItems: "baseline",
-                  flexWrap: "wrap",
-                }}
-              >
-                <Text
-                  style={{
-                    color: theme.semantic.text.quaternary,
-                    fontSize: statFontSize,
-                    lineHeight: statFontSize * 1.2,
-                    marginRight: 2,
-                  }}
-                >
-                  {"Number of Hints Used: "}
-                </Text>
-                <View style={{ flexDirection: "row", alignItems: "center" }}>
-                  <Text
-                    style={{
-                      color: theme.semantic.text.primary,
-                      fontSize: statFontSize,
-                      lineHeight: statFontSize * 1.2,
-                      fontWeight: "bold",
-                    }}
-                    testID="numHintsUsed"
-                  >
-                    {statistics.numHintsUsed}
-                  </Text>
-                  <List.Icon
-                    icon={
-                      isHintsBreakdownExpanded ? "chevron-up" : "chevron-down"
-                    }
-                    color={theme.semantic.text.primary}
-                    style={{ margin: 0 }}
-                  />
-                </View>
-              </View>
-            </TouchableRipple>
-            {isHintsBreakdownExpanded ? (
-              <NumHintsUsedPerStrategy
-                numHintsUsedPerStrategy={statistics.numHintsUsedPerStrategy}
-              />
-            ) : null}
+            <ExpandableHintsUsedStatistic
+              numHintsUsed={statistics.numHintsUsed}
+              numHintsUsedPerStrategy={statistics.numHintsUsedPerStrategy}
+            />
           </ScrollView>
         </View>
